moving other functions(THS)
